git2graph: add (static) support for filtering desired nodes & edges
This allows to select desired nodes and edges based on their types, similarly
to what the edges restriction field in the swh-graph REST API does. (It is in
fact strictly more expressive, as one can also select nodes without the
corresponding edges.)

The support is only static for now: boolean arrays (or matrices) have to be
changed in the source code to alter the selection, which by default is "all".
Support for making this user-selectable on the CLI is upcoming.