binary maps: change type IDs on Python side, to be compatible with Java
Before this change we had an off-by-1: Java type integer IDs were 0-based,
Python ones 1-based. With this change they match and are both 0-based.

WARNING: with this change we break backward compatibility for the Python client
when reading binary maps that were generated (via a Python hack) before this
change. They will need to be regenerated either using the now available
Java-based generation of binary maps or by rerunning the Python hack with the
new code.