FindEarliestRevision: bug fix: do not follow rev:rev edges
Rationale: we do not need to do so at all, because we are only interested in
commits that directly contain the content or dir (at some depth), nor in any of
their successors.

Note: this bug was not responsible for wrong answers in most cases (because
successors will tend to have higher timestamps), but incurred significant extra
time (4x in early benchmarks) due to exploring significant commit histories for
any file/dir appearing in large projects.