Delay import to allow proper debian packaging of the client code
The swh.graph.client and swh.graph.naive_client must be packaged separately to allow the
vault to use the optional graph dependency. We need the tests to run in the packaging
though. The current conftest.py mixes both server and client code though. Delaying
imports this way will allow the packaging to run accordingly [1].

[1]
```
pytest -k "not remote" swh/
```

Related to T3518