Move SWHID<->node ID conversion in the Java backend
Doing the SWHID <-> node ID conversion in the Python side prevents using
the minimal perfect hash function for SWHID -> node ID. This design is
inherited from when we thought we could have a really thin Java layer
and write all the code in Python, but this turned out to be infeasible
in practice.

Instead, we are instead moving more and more things in the Java layer,
until ultimately the entire Python RPC API might be replaced by a pure
Java implementation.

This commit drops the swh/graph/graph.py file for the same reasons
outlined above (it's infeasible to write graph traversal algorithms in
pure Python), along with its associated tests.