Add graph properties compressed from the ORC dataset
This commit adds the handling of graph *properties*, i.e., data attached
to nodes or edges (commit timestamps, commit messages, content lengths,
...) to swh-graph.

The class WriteNodeProperties is used to extract the node properties
from the ORCGraphDataset and write them in separate files, in compressed
format. The properties can then be read using the SwhGraphProperties
class.

The compression pipeline and the tests were all changed to use the new
dataset format.