Bump requirements on protobuf
protobuf earlier 4.12.* versions crashed when `swhgraph_pb2.py`
was discovered by pytest. This has been fixed in the 4.12.11 release.
See: https://github.com/protocolbuffers/protobuf/issues/10151

Preventing protobuf 4.12.* to be used makes `pip` install grpcio-tools
version 1.49.0 which fails to build on Debian bookworm. In order to
allow more fixed versions of grpcio-tools to be used, we bump the
dependency on protobuf to version 4.12.11 or later.