Add export_{started,ended}_at to Stats response
After a graph is exported, Software Heritage archive will
continue to grow and change. Currently clients have no easy way
to know when was made the export available from a swh.graph server.
Thus making it harder to figure out was has been added to swh.storage
since.

Therefore, we add a new fields to the response of the Stats API,
“export_started_at” and “export_ended_at”. These fields are in “seconds
since epoch” like other time values currently in the protocol. They are
marked as optional as the required data might be missing or currupted.

The values are taken from the `meta/export.json` file. The `meta`
directory needs to be in the same directory where webgraph data
lies.

Closes #4779