Move the doc for the example dataset to its own page
Having the documentation of the example dataset inside the
`swh.graph.example_dataset` module sadly prevent to include
the visualization in a way that works with building both swh-graph and
swh-docs.

To fix this issue, we move the documentation of the example dataset to
its own page in `docs/`. As the example dataset can be used entirely
without the Python module, it is probably for the best.