Raise an error when RemoteGraphClient URL is wrong
We used to allow users to create a RemoteGraphClient with any URL.
This gets surprising results if the HTTP server is up, but the location
is wrong. As the HTTP Graph API will return 404 for unknown SWHIDs,
a client can fail to notice the configuration problem and silently get
wrong results.

Instead, we now query the `/stats` endpoint on instanciation and raise
exceptions in case of 404, connection error or unexpected content in
stats results.

Closes #4789