Fix LongArrayBitVector to work with graphs over 8000 nodes
I only tested on a small bit vector and assumed longer ones would have the
same structure, but they don't.

Java writes them in small chunks, separated by 5 bytes each, which makes
it impossible to use with NumberMmap.

Instead, LongArrayBitVector will now load the whole file in memory like
the Java class instead of mmapping. We should probably change this in the
future, to either mmap it and work around the weird format, or change
the on-disk format so LongArrayBitVector isn't needed anymore.