[bad] Parallelize list-by-earliest-revision
At first it looked like a good idea because it improves the throughput by 10x
when .node2swhid and .order are not fully loaded in memory.

But after they are loaded, it is a little slower while using 6 times as much
CPU time. Not to mention the extra complicated code.