provenance: Switch from glibc malloc to jemalloc
The three executables building the final provenance index
(contents-in-directories, contents-in-revisions-without-frontier, and
frontier-directories-in-revisions) make lots of small allocations when
concatenating paths.

With glibc malloc on a 96 threads system, this was causing 1/3 to 2/3rd
of the time in a lock.

Switching to jemalloc brings a 6x speedup overall.

While not strictly needed, this commit switches other provenance
executables to jemalloc as well, for consistency (so profilings aren't
inconsistent from one to the other).