Move gRPC server to its own crate
This is easier to install, as there is no need for users to remember
--features=grpc-server.

This also avoids declaring every dependency of the gRPC server as 'optional' and
adding it to the list of dependencies of the 'grpc-server' feature; and I expect
this set of dependencies to get larger now that the gRPC server is starting to
be used in production (eg. statsd metrics).

Finally, this allows Cargo to remove building the gRPC server from the critical
path to building swh_graph_provenance when building all crates (eg. in CI),
which should make it a little faster to build.