docs: Recommend using --locked with cargo install
The --locked option forces Cargo to use the lockfile declared by swh-graph,
which was carefully written to avoid conflicts between two Arrow versions
that are used as transitive dependencies through different packages.

When Cargo tries to update one, it ends up with incompatible versions,
causing a compilation failure.

Currently, this only happens when enabling the `orc` features (which
commands in the documentation do not enable), but similar issues could
appear in the future with the default versions.