Fix cgroup management when calling .run() twice on the same Command
The cgroup was created when the Command is initialized, and calling .run() on it
twice caused creation of two RunningCommand objects.

As RunningCommand is responsible for deleting the cgroup (and polling information
from it), this means that one of the RunningCommand deleted the cgroup before the
other one tried to read it, causing a crash.

This happens for example in CountPaths when direction=backward, because the
zstdcat command running on the topological order is used twice (once to read the
header, once to read the rest of the file)