Add support for making properties optional at runtime
`SwhGraphProperties` takes five different type parameters, to indicate which properties are loaded, and this is defined at compile time. Usually this is fine, because a given script's author knows at compile time which properties it needs to access, and which it doesn't. However, the gRPC server (and possible some other nice cases) **may** want to access any of these properties, but it depends on users, at runtime, which of them they will access. As some of these properties are backed by very large file, we would like to make them optional at runtime.

The naive design would be to just remove the type parameters, and make every property always runtime-optional. However, this has downsides:

1. swh-graph needs to check the `Option<Mmap>` every time it accesses a property (which has some code size and possibly runtime overhead)
2. users need to handle the `None` case every time they access a property (which complicates their code)
3. it breaks backward compatibility

A less naive design would be to duplicate every type parameter and every property accessor, but this means we need to duplicate the code of every accessor; and it prevents authors of libraries based on swh-graph from "forwarding" this conditionality to their own users, without also duplicating their code. (And it is not possible to deduplicate without using macros, as the "always available property" and "runtime-optional property" would be completely different types.)

Instead, this MR refactors the type and trait system used as type parameters to `SwhGraphProperties`, by adding a new trait (eg. `OptTimestamps`) that expresses a set of properties may be available at runtime (eg. as opposed to `Timestamps` which is unconditional) and new structures (eg. `OptMappedStrings` in addition to `MappedStrings`). Accessor code is then made generic over both variants, and so can user code (so eg. this allows a single codebase for the gRPC server, one with the conditionals and an hypothetical one without to avoid the overhead if it turns out to be an issue). For example, this code:

```
    pub fn author_timestamp_offset(&self, node_id: NodeId) -> Option<i16> {
        self.try_author_timestamp_offset(node_id)
            .unwrap_or_else(|e| panic!("Cannot get author timestamp offset: {}", e))
```

is replaced with:

```
    pub fn author_timestamp_offset(
        &self,
        node_id: NodeId,
    ) -> PropertiesResult<Option<i16>, TIMESTAMPS> {
        TIMESTAMPS::map_if_available(
            self.try_author_timestamp_offset(node_id),
            |author_timestamp_offset| {
                author_timestamp_offset
                    .unwrap_or_else(|e| panic!("Cannot get author timestamp offset: {}", e))
            },
        )
    }
```

which is monomorphized to the same code in the non-optional case, but is monomorphized to something like this in the optional case:

```
    pub fn author_timestamp_offset(
        &self,
        node_id: NodeId,
    ) -> Result<Option<i16>, UnavailableProperty> {
        Ok(self.try_author_timestamp_offset(node_id)?
            .unwrap_or_else(|e| panic!("Cannot get author timestamp offset: {}", e)))
    }
```

For most users, this is transparent, as their existing code is not affected.