pytest_plugin: Ensure to kill swh-graph-grpc-serve process on teardown
This process is spawned by the graph HTTP server process but this latter
is killed when tearing down the graph_grpc_server_process session scope
fixture and thus the spawned swh-graph-grpc-serve process is still
running afterwards.

Ensure to kill the gRPC server process by sending a SIGKILL signal to
it after killing the graph HTTP server.