pytest_plugin: Use safer multiprocessing start method for graph server
According to Python multiprocessing documentation, safely forking a
multithreaded process is problematic.

This assertion can be verified when running the swh-graph Python tests
as sometimes a "Fatal Python error: Aborted" is raised when attempting
to create the graph server process through the graph_grpc_server_process
fixture.

Prefer to use the "forkserver" start method of multiprocessing that is
safer to use when creating processes from a multithreaded process.