Refactor SwhidMphf to embed a permutation
Before this commit, SwhidMphf only represented 'naked' MPHFs, and
src/properties/maps.rs applied a permutation after calling the MPHF

However, some MPHFs are actually static functions, ie. they
intrinsically perform the role of both a perfect-hash function and a
permutation. For these, it is then wasteful to have a permutation next
to it.

This is already the case of VecMaps, which allocated an identity
permutation. This was not a big deal because VecMaps is only used in
tests or in small graphs.

However, we are considering switching from PTHash to VFunc as the MPH
we use in real graphs, and VFunc is a static function. This means that
for new graphs, we should not load nor use a permutation.

The naive implementation of this would be to make the permutation an
`Option<MappedPermutation>`, but this requires a runtime check in a hot
loop (that is already a bottleneck in some cases), so instead this
relies on typing to eliminate the runtime check at compile time.

This is technically a breaking change, as it is possible for user code
to be generic on `SwhidMphf`, but very unlikely.