example_dataset: Update some edge labels and regenerate dataset
Previously the set of edge labels (directory entry names and snapshot branch
names) in the graph example dataset could lead to an infinite loop of the
swh-graph-compress process when running the tests. The issue can happen when
building the MPH for labels due to a bug in the pthash C++ library used under
the hood by pthash-rs.

Modify two directory entry names and regenerate the example dataset to avoid
hitting the infinite loop issue and ensure stability of the swh-graph tests.

Update Python and Rust tests according to these changes.

Fixes #4857.