Refactor properties backends to provide indexing themselves
Returning a structure to index in makes it harder to provide alternative implementations; typically structures that are not slice-like enough.

This will be used to provide ContiguousSubgraph (https://gitlab.softwareheritage.org/swh/devel/swh-graph/-/merge_requests/794), which exposes nodes in a way that two consecutive nodes in the ContiguousSubgraph are not necessarily adjacent in the underlying graph (so, also not in Node2SWHID/Node2Type).

This would also be needed by the hypothetical "union graph" that would implement updates to an existing compressed graph (https://gitlab.softwareheritage.org/swh/devel/swh-graph/-/issues/4832).