download: Implement resume download feature in GraphDownloader class
Previously if an error happened when downloading a file of a graph dataset
an exception was raised leaving the downloaded dataset in an incomplete and
unusable state.

Moreover, when attempting to re-download the dataset, all files were entirely
downloaded again regardless if they were already successfully or partially
downloaded, which can be really cumbersome when the full dataset to download
weights several terabytes.

To prevent such annoying download behavior, update the GraphDownloader class
implementation by adding a resume download feature. It allows to keep downloaded
dataset in a partial state in order to fetch only the missing files or the missing
bytes of downloads that previously fail when attempting a new dataset download.

Replace use of multiprocessing.dummy by concurrent.futures.

Due to these changes, downloaded files compressed with zstd are now uncompressed
after the full .zst files have been downloaded.

Fixes #4886.